+++ tools/SetupFlow/DevHome.SetupFlow/DevHome.SetupFlow.csproj
@@ -32,16 +32,25 @@
     <None Update="Assets\AppManagementPackages.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <Page Update="Views\Summary\SummaryFailedTasks.xaml">
+    <Page Update="Views\Summary\SummaryAppInstallationNotes.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Update="Views\Summary\SummaryShowAppsAndRepos.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
+    <Page Update="Views\Summary\SummaryAppsDownloadedReposCloned.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Update="Views\Summary\SummaryFailedTasks.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Update="Views\Summary\SummaryTargetMachineShowAppsAndRepos.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Update="Views\Summary\SummaryNextSteps.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Update="Views\Summary\SummaryAppsDownloadedReposCloned.xaml">
+    <Page Update="Views\Summary\SummaryTargetMachineAppsDownloadedReposCloned.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Update="Views\Summary\SummaryTargetMachineWithErrors.xaml">
@@ -123,6 +132,7 @@
     <None Remove="Views\SetupTargetReviewView.xaml" />
     <None Remove="Views\SetupTargetView.xaml" />
     <None Remove="Views\SummaryView.xaml" />
+    <None Remove="Views\Summary\SummaryAppInstallationNotes.xaml" />
     <None Remove="Views\Summary\SummaryAppsDownloadedReposCloned.xaml" />
     <None Remove="Views\Summary\SummaryConfigurationFileResults.xaml" />
     <None Remove="Views\Summary\SummaryFailedTasks.xaml" />
@@ -131,6 +141,8 @@
     <None Remove="Views\Summary\SummaryNeedsRestart.xaml" />
     <None Remove="Views\Summary\SummaryNextSteps.xaml" />
     <None Remove="Views\Summary\SummaryShowAppsAndRepos.xaml" />
+    <None Remove="Views\Summary\SummaryTargetMachineAppsDownloadedReposCloned.xaml" />
+    <None Remove="Views\Summary\SummaryTargetMachineShowAppsAndRepos.xaml" />
     <None Remove="Views\Summary\SummaryTargetMachineWithErrors.xaml" />
     <None Remove="Windows\DevDriveWindow.xaml" />
     <None Remove="Windows\InstallationNotesWindow.xaml" />
