+++ tools/SetupFlow/DevHome.SetupFlow/Views/Summary/SummaryAppsDownloadedReposCloned.xaml
@@ -19,18 +19,19 @@
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </UserControl.Resources>
-    <Grid>
+    <Grid BorderBrush="Red" BorderThickness="1" VerticalAlignment="Top">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="{Binding AppsDownloaded, Converter={StaticResource EmptyCollectionWillNotSpanColumnsConverter}}"/>
             <ColumnDefinition Width="{Binding RepositoriesCloned, Converter={StaticResource EmptyCollectionWillNotSpanColumnsConverter}}"/>
             <ColumnDefinition />
         </Grid.ColumnDefinitions>
         <Grid.RowDefinitions>
-            <RowDefinition />
-            <RowDefinition />
-            <RowDefinition />
+            <RowDefinition Height="auto"/>
+            <RowDefinition Height="auto"/>
         </Grid.RowDefinitions>
         <Grid
+                        BorderBrush="blue"
+            BorderThickness="10"
         Visibility="{Binding AppsDownloaded, Converter={StaticResource CollectionVisibilityConverter}}"
         Grid.Column="0"
         HorizontalAlignment="Center">
@@ -51,6 +52,8 @@
         </Grid>
         <Grid
     Grid.Column="1"
+                        BorderBrush="blue"
+            BorderThickness="10"
     Visibility="{Binding RepositoriesCloned, Converter={StaticResource CollectionVisibilityConverter}}">
             <Grid.RowDefinitions>
                 <RowDefinition/>
@@ -93,6 +96,8 @@
         HorizontalAlignment="Center" 
         Command="{Binding RedirectToNextPageCommand, Mode=OneWay}" 
         Style="{StaticResource AccentButtonStyle}"
+                BorderBrush="Green"
+                BorderThickness="10"
         Margin="0, 50, 0, 0"/>
     </Grid>
 </UserControl>
