+++ tools/SetupFlow/DevHome.SetupFlow/Views/Summary/SummaryFailedTasks.xaml.cs
@@ -1,28 +1,14 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Microsoft.UI.Xaml;
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using Microsoft.UI.Xaml.Data;
-using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Navigation;
 
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
+namespace DevHome.SetupFlow.Views.Summary;
 
-namespace DevHome.SetupFlow.Views.Summary
+public sealed partial class SummaryFailedTasks : UserControl
 {
-    public sealed partial class SummaryFailedTasks : UserControl
+    public SummaryFailedTasks()
     {
-        public SummaryFailedTasks()
-        {
-            this.InitializeComponent();
-        }
+        this.InitializeComponent();
     }
 }
