+++ tools/SetupFlow/DevHome.SetupFlow/Views/Summary/SummaryShowAppsAndRepos.xaml.cs
@@ -1,28 +1,67 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AdaptiveCards.Rendering.WinUI3;
+using CommunityToolkit.Mvvm.Messaging;
+using DevHome.SetupFlow.Models.Environments;
+using DevHome.SetupFlow.ViewModels;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using Microsoft.UI.Xaml.Data;
-using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Navigation;
 
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
+namespace DevHome.SetupFlow.Views.Summary;
 
-namespace DevHome.SetupFlow.Views.Summary
+public sealed partial class SummaryShowAppsAndRepos : UserControl, IRecipient<NewAdaptiveCardAvailableMessage>
 {
-    public sealed partial class SummaryShowAppsAndRepos : UserControl
+    public SummaryShowAppsAndRepos()
     {
-        public SummaryShowAppsAndRepos()
+        this.InitializeComponent();
+        WeakReferenceMessenger.Default.Register<NewAdaptiveCardAvailableMessage>(this);
+    }
+
+    /// <summary>
+    /// Receive the adaptive card from the view model, when the view model finishes loading it.
+    /// Note: There are times when the view is loaded after the view model has finished loading the adaptive card.
+    /// In these cases it would have "missed" the push message. This is where the ViewLoaded method comes in.
+    /// </summary>
+    public void Receive(NewAdaptiveCardAvailableMessage message)
+    {
+        // Only process the message if the view model is the SummaryViewModel
+        if (message.Value.CurrentSetupFlowViewModel is SummaryViewModel)
         {
-            this.InitializeComponent();
+            AddAdaptiveCardToUI(message.Value.RenderedAdaptiveCard);
         }
+    }
+
+    /// <summary>
+    /// Request the adaptive cad from the EnvironmentCreationOptionsViewModel object when we're in the environment
+    /// creation flow.
+    /// </summary>
+    private void ViewLoaded(object sender, RoutedEventArgs e)
+    {
+        var message = WeakReferenceMessenger.Default.Send<CreationOptionsReviewPageDataRequestMessage>();
+        if (!message.HasReceivedResponse)
+        {
+            return;
+        }
+
+        AddAdaptiveCardToUI(message.Response);
+    }
+
+    private void ViewUnloaded(object sender, RoutedEventArgs e)
+    {
+        AdaptiveCardGrid.Children.Clear();
+        WeakReferenceMessenger.Default.UnregisterAll(this);
+    }
+
+    private void AddAdaptiveCardToUI(RenderedAdaptiveCard renderedAdaptiveCard)
+    {
+        var frameworkElement = renderedAdaptiveCard?.FrameworkElement;
+        if (frameworkElement == null)
+        {
+            return;
+        }
+
+        AdaptiveCardGrid.Children.Clear();
+        AdaptiveCardGrid.Children.Add(frameworkElement);
     }
 }
